Update README.md
As a skilled software engineer with a deep focus on machine learning, AI, and backend development, I bring a unique combination of technical expertise and problem-solving abilities to every project. With advanced degrees in Software Engineering and a specialization in Deep Learning and Computer Vision, I have a strong foundation in developing and deploying cutting-edge AI solutions.